requirements.txt
alembic==1.13.3
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
asttokens==2.4.1
asyncpg==0.29.0
attrs==24.2.0
autoflake==2.3.1
Automat==24.8.1
bcrypt==3.2.0
billiard==4.2.1
black==24.10.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cors==1.0.1
debugpy==1.8.7
decorator==5.1.1
dnspython==2.7.0
docker==7.1.0
ecdsa==0.19.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.4
fastapi-cache2==0.2.2
fastapi-versioning==0.10.0
filelock==3.16.1
flake8==7.1.1
flower==2.0.1
future==1.0.0
gevent==24.10.3
grafana==0.0.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
humanize==4.11.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.28.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
kombu==5.4.2
Mako==1.3.5
MarkupSafe==3.0.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
netcat==0.0.1a0
nodeenv==1.9.1
outcome==1.3.0.post0
packaging==24.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.18.0
pyright==1.1.388
PySocks==1.7.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-tornasync==0.6.0.post2
pytest-trio==0.8.0
pytest-twisted==1.14.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.17
pytz==2024.2
pywin32==308; platform_system=="Windows"
pyzmq==26.2.0
redis==5.2.0
requests==2.32.3
requests-file==2.1.0
rsa==4.9
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqladmin==0.20.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.2
time-machine==2.16.0
tldextract==5.1.2
tornado==6.4.1
traitlets==5.14.3
trio==0.27.0
Twisted==24.10.0
typing_extensions==4.12.2
tzdata==2024.2
unicorn==2.1.1
urllib3==2.2.3
uvicorn==0.32.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.0
WTForms==3.1.2
zope.event==5.0
zope.interface==7.1.1
